refactor(CrossAppLogin): Update ExternalAccount prior to 1st public release
Use a Long for the id, to not have to break compatibility if it ever grows.
Reorder properties, and change the proto numbers to be ordered while we can.

+++ CrossAppLoginUI/src/main/kotlin/com/infomaniak/core/crossloginui/views/components/CrossLoginSelectAccounts.kt
@@ -40,7 +40,7 @@ import com.infomaniak.core.login.crossapp.ExternalAccount
 @Composable
 fun CrossLoginSelectAccounts(
     accounts: SnapshotStateList<ExternalAccount>,
-    skippedIds: SnapshotStateSet<Int>,
+    skippedIds: SnapshotStateSet<Long>,
     customization: CrossLoginCustomization = CrossLoginDefaults.customize(),
     onClick: () -> Unit,
 ) {
@@ -65,7 +65,7 @@ private fun Preview(@PreviewParameter(AccountsPreviewParameter::class) accounts:
     Surface {
         CrossLoginSelectAccounts(
             accounts = remember { mutableStateListOf<ExternalAccount>().apply { addAll(accounts) } },
-            skippedIds = remember { mutableStateSetOf<Int>().apply { add(accounts.last().id) } },
+            skippedIds = remember { mutableStateSetOf<Long>().apply { add(accounts.last().id) } },
             onClick = {},
         )
     }
